@@ -1,23 +1,41 @@
 package nightkosh.gravestone.core.event;
 
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.passive.*;
-import net.minecraft.entity.player.EntityPlayer;
-import net.minecraftforge.event.entity.EntityJoinWorldEvent;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.world.damagesource.DamageTypes;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.TamableAnimal;
+import net.minecraft.world.entity.animal.Cat;
+import net.minecraft.world.entity.animal.Wolf;
+import net.minecraft.world.entity.animal.horse.AbstractHorse;
+import net.minecraft.world.entity.animal.horse.Horse;
+import net.minecraft.world.entity.npc.Villager;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.storage.LevelResource;
+import net.minecraftforge.event.AttachCapabilitiesEvent;
+import net.minecraftforge.event.RegisterCommandsEvent;
+import net.minecraftforge.event.entity.EntityJoinLevelEvent;
+import net.minecraftforge.event.entity.EntityLeaveLevelEvent;
 import net.minecraftforge.event.entity.living.LivingDeathEvent;
-import net.minecraftforge.event.entity.player.PlayerDropsEvent;
+import net.minecraftforge.event.entity.living.LivingDropsEvent;
 import net.minecraftforge.event.entity.player.PlayerEvent;
-import net.minecraftforge.event.world.WorldEvent;
-import net.minecraftforge.fml.common.FMLCommonHandler;
-import net.minecraftforge.fml.common.eventhandler.EventPriority;
-import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
-import nightkosh.gravestone.api.death_handler.*;
-import nightkosh.gravestone.config.Config;
+import net.minecraftforge.event.level.LevelEvent;
+import net.minecraftforge.eventbus.api.EventPriority;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.fml.common.Mod;
+import nightkosh.gravestone.api.ModInfo;
+import nightkosh.gravestone.capability.BackupProvider;
+import nightkosh.gravestone.config.GSConfigs;
+import nightkosh.gravestone.core.GSCommands;
 import nightkosh.gravestone.core.MobHandler;
 import nightkosh.gravestone.core.logger.GravesLogger;
 import nightkosh.gravestone.helper.BackupsHelper;
 import nightkosh.gravestone.helper.GraveGenerationHelper;
 import nightkosh.gravestone.helper.api.APIGraveGeneration;
+
+import static nightkosh.gravestone.ModGraveStone.LOGGER;
 
 /**
  * GraveStone mod
@@ -25,56 +43,87 @@ import nightkosh.gravestone.helper.api.APIGraveGeneration;
  * @author NightKosh
  * @license Lesser GNU Public License v3 (http://www.gnu.org/licenses/lgpl.html)
  */
+@Mod.EventBusSubscriber(modid = ModInfo.ID, bus = Mod.EventBusSubscriber.Bus.FORGE)
 public class EventsHandler {
 
+    private static final ResourceLocation BACKUPS_RL = new ResourceLocation(ModInfo.ID, "backups");
+
+    @SubscribeEvent
+    public static void onAttachCaps(AttachCapabilitiesEvent<Entity> event) {
+        if (event.getObject() instanceof Player player) {
+            if (GSConfigs.DEBUG_MODE.get()) {
+                LOGGER.info("AttachCapabilitiesEvent event triggered for player {}", player.getUUID().toString());
+            }
+            event.addCapability(BACKUPS_RL, new BackupProvider());
+        }
+    }
+
     @SubscribeEvent(priority = EventPriority.HIGHEST)
-    public void onPlayerClone(PlayerEvent.Clone event) {
+    public static void onPlayerClone(PlayerEvent.Clone event) {
         if (event.isWasDeath()) {
-            BackupsHelper.clonePlayer(event.getOriginal(), event.getEntityPlayer());
+            if (GSConfigs.DEBUG_MODE.get()) {
+                LOGGER.info("PlayerEvent.Clone event triggered for {}", event.getEntity().getScoreboardName());
+            }
+            BackupsHelper.clonePlayer(event.getOriginal(), event.getEntity());
         }
     }
 
     @SubscribeEvent(priority = EventPriority.HIGH)
-    public void onEntityLivingDeath(LivingDeathEvent event) {
-        if (FMLCommonHandler.instance().getEffectiveSide().isServer()) {
-            if (!Config.generateGravesInLava && event.getSource().damageType.equals("lava")) {
+    public static void onEntityLivingDeath(LivingDeathEvent event) {
+        if (!event.getEntity().level.isClientSide()) {
+            if (GSConfigs.DEBUG_MODE.get()) {
+                LOGGER.info("LivingDeathEvent event triggered");
+            }
+
+            if (!GSConfigs.GENERATE_GRAVES_IN_LAVA.get() && event.getSource().is(DamageTypes.LAVA)) {
+                if (GSConfigs.DEBUG_MODE.get()) {
+                    LOGGER.info("Event skipped - death in lava");
+                }
                 return;
             }
 
-            if (Config.generateVillagerGraves && event.getEntity() instanceof EntityVillager) {
-                EntityVillager villager = (EntityVillager) event.getEntity();
-                for (IVillagerDeathHandler villagerDeathHandler : APIGraveGeneration.VILLAGER_DEATH_HANDLERS) {
+            if (GSConfigs.GENERATE_VILLAGER_GRAVES.get() && event.getEntity() instanceof Villager villager) {
+                for (var villagerDeathHandler : APIGraveGeneration.VILLAGER_DEATH_HANDLERS) {
                     if (villagerDeathHandler.cancelGraveGeneration(villager, event.getSource())) {
+                        if (GSConfigs.DEBUG_MODE.get()) {
+                            LOGGER.info("Villager grave generation cancelled");
+                        }
                         return;
                     }
                 }
                 GraveGenerationHelper.createVillagerGrave(villager, event.getSource());
                 return;
-            } else if (Config.generatePetGraves) {
-                if (event.getEntity() instanceof EntityTameable) {
-                    if (event.getEntity() instanceof EntityWolf) {
-                        EntityWolf dog = (EntityWolf) event.getEntity();
-                        for (IDogDeathHandler dogDeathHandler : APIGraveGeneration.DOG_DEATH_HANDLERS) {
+            } else if (GSConfigs.GENERATE_PET_GRAVES.get()) {
+                if (event.getEntity() instanceof TamableAnimal) {
+                    if (event.getEntity() instanceof Wolf dog) {
+                        for (var dogDeathHandler : APIGraveGeneration.DOG_DEATH_HANDLERS) {
                             if (dogDeathHandler.cancelGraveGeneration(dog, event.getSource())) {
+                                if (GSConfigs.DEBUG_MODE.get()) {
+                                    LOGGER.info("Dog grave generation cancelled");
+                                }
                                 return;
                             }
                         }
                         GraveGenerationHelper.createDogGrave(dog, event.getSource());
                         return;
-                    } else if (event.getEntity() instanceof EntityOcelot) {
-                        EntityOcelot cat = (EntityOcelot) event.getEntity();
-                        for (ICatDeathHandler catDeathHandler : APIGraveGeneration.CAT_DEATH_HANDLERS) {
+                    } else if (event.getEntity() instanceof Cat cat) {
+                        for (var catDeathHandler : APIGraveGeneration.CAT_DEATH_HANDLERS) {
                             if (catDeathHandler.cancelGraveGeneration(cat, event.getSource())) {
+                                if (GSConfigs.DEBUG_MODE.get()) {
+                                    LOGGER.info("Cat grave generation cancelled");
+                                }
                                 return;
                             }
                         }
                         GraveGenerationHelper.createCatGrave(cat, event.getSource());
                         return;
                     }
-                } else if (event.getEntity() instanceof AbstractHorse) {
-                    AbstractHorse horse = (AbstractHorse) event.getEntity();
-                    for (IHorseDeathHandler horseDeathHandler : APIGraveGeneration.HORSE_DEATH_HANDLERS) {
+                } else if (event.getEntity() instanceof AbstractHorse horse) {
+                    for (var horseDeathHandler : APIGraveGeneration.HORSE_DEATH_HANDLERS) {
                         if (horseDeathHandler.cancelGraveGeneration(horse, event.getSource())) {
+                            if (GSConfigs.DEBUG_MODE.get()) {
+                                LOGGER.info("Horse grave generation cancelled");
+                            }
                             return;
                         }
                     }
@@ -83,7 +132,7 @@ public class EventsHandler {
                 }
             }
 
-            for (ICustomEntityDeathHandler customEntityDeathHandler : APIGraveGeneration.CUSTOM_ENTITY_DEATH_HANDLERS) {
+            for (var customEntityDeathHandler : APIGraveGeneration.CUSTOM_ENTITY_DEATH_HANDLERS) {
                 if (event.getEntity().getClass().equals(customEntityDeathHandler.getEntityClass()) &&
                         customEntityDeathHandler.canGenerateGrave(event.getEntity(), event.getSource())) {
                     GraveGenerationHelper.createCustomGrave(event.getEntity(), event, customEntityDeathHandler);
@@ -93,61 +142,115 @@ public class EventsHandler {
         }
     }
 
-
     @SubscribeEvent(priority = EventPriority.LOWEST)
-    public void onPlayerDrops(PlayerDropsEvent event) {
-        if (FMLCommonHandler.instance().getEffectiveSide().isServer()) {
-            if (!Config.generateGravesInLava && event.getSource().damageType.equals("lava")) {
+    public static void onPlayerDrops(LivingDropsEvent event) {
+        if (!event.getEntity().level.isClientSide()) {
+            if (GSConfigs.DEBUG_MODE.get()) {
+                LOGGER.info("LivingDropsEvent event triggered");
+            }
+
+            if (!GSConfigs.GENERATE_GRAVES_IN_LAVA.get() && event.getSource().is(DamageTypes.LAVA)) {
+                if (GSConfigs.DEBUG_MODE.get()) {
+                    LOGGER.info("Event skipped - death in lava");
+                }
                 return;
             }
 
-            if (Config.generatePlayerGraves && event.getEntityLiving() instanceof EntityPlayer) {
-                EntityPlayer player = (EntityPlayer) event.getEntity();
-                if (!Config.playerGravesDimensionalBlackList.contains(player.dimension)) {
-                    for (IPlayerDeathHandler playerDeathHandler : APIGraveGeneration.PLAYER_DEATH_HANDLERS) {
+            if (GSConfigs.GENERATE_PLAYER_GRAVES.get() && event.getEntity() instanceof Player player) {
+                if (!GSConfigs.PLAYER_GRAVES_DIMENSIONAL_BLACKLIST.get().contains(
+                        player.level.dimension().location().toString())) {
+                    for (var playerDeathHandler : APIGraveGeneration.PLAYER_DEATH_HANDLERS) {
                         if (playerDeathHandler.cancelGraveGeneration(player, event.getSource())) {
+                            if (GSConfigs.DEBUG_MODE.get()) {
+                                LOGGER.info("Player {} grave generation cancelled", player.getScoreboardName());
+                            }
                             return;
                         }
                     }
-
                     GraveGenerationHelper.createPlayerGrave(player, event.getDrops(), event.getSource(), MobHandler.getAndRemoveSpawnTime(player));
+                } else if (GSConfigs.DEBUG_MODE.get()) {
+                    LOGGER.info("Player {} grave generation cancelled in blacklisted dimension", player.getScoreboardName());
                 }
             }
         }
     }
 
     @SubscribeEvent
-    public void entityJoinWorldEvent(EntityJoinWorldEvent event) {
-        if (FMLCommonHandler.instance().getEffectiveSide().isServer()) {
-            Entity entity = event.getEntity();
-            if (entity instanceof EntityVillager ||
-                    entity instanceof EntityWolf ||
-                    entity instanceof EntityOcelot ||
-                    entity instanceof EntityHorse) {
-                MobHandler.setMobSpawnTime(event.getEntity());
+    public static void entityJoinWorldEvent(EntityJoinLevelEvent event) {
+        if (!event.getEntity().level.isClientSide() && event.getEntity() instanceof LivingEntity entity) {
+            if (entity instanceof Villager ||
+                    entity instanceof Wolf ||
+                    entity instanceof Cat ||
+                    entity instanceof Horse) {
+                if (GSConfigs.DEBUG_MODE.get()) {
+                    LOGGER.info("EntityJoinLevelEvent event triggered for villager, wolf, cat or horse ({})",
+                            entity.getScoreboardName());
+                }
+                MobHandler.setMobSpawnTime(entity);
             }
         }
     }
 
-    // TODO remove mobs info at despawn
-
     @SubscribeEvent
-    public void worldLoading(WorldEvent.Load event) {
-        if (FMLCommonHandler.instance().getEffectiveSide().isServer()) {
-            MobHandler.loadMobsSpawnTime(event.getWorld());
-            GravesLogger.setWorldDirectory(event.getWorld().getSaveHandler().getWorldDirectory());
+    public static void entityLeaveLevelEvent(EntityLeaveLevelEvent event) {
+        if (!event.getEntity().level.isClientSide() && event.getEntity() instanceof LivingEntity entity) {
+            if (!entity.isAlive() &&
+                    entity instanceof Villager ||
+                    entity instanceof Wolf ||
+                    entity instanceof Cat ||
+                    entity instanceof Horse) {
+                if (GSConfigs.DEBUG_MODE.get()) {
+                    LOGGER.info("EntityLeaveLevelEvent event triggered for villager, wolf, cat or horse ({})",
+                            entity.getScoreboardName());
+                }
+                MobHandler.removeSpawnTime(entity);
+            }
         }
     }
 
-    //TODO remove #245 !!!!!!!!!!!!!!!!!!!!!!!
-//    @SubscribeEvent
-//    public void onChunkLoad(ChunkEvent.Load event) {
-//        if (FMLCommonHandler.instance().getEffectiveSide().isServer()) {
-//            Chunk chunk = event.getChunk();
-//            chunk.getTileEntityMap().entrySet().stream().filter(entry -> entry.getValue() instanceof TileEntityGrave).forEach(entry -> {
-//                GraveStoneHelper.replaceOldGraveByNew(event.getWorld(), entry.getKey());
-//            });
-//        }
-//
-//    }
+    @SubscribeEvent
+    public static void onLevelLoad(LevelEvent.Load event) {
+        if (event.getLevel() instanceof ServerLevel serverLevel &&
+                serverLevel.dimension() == Level.OVERWORLD &&
+                !serverLevel.isClientSide()) {
+            if (GSConfigs.DEBUG_MODE.get()) {
+                LOGGER.info("LevelEvent.Load event triggered");
+            }
+            var worldPath = serverLevel.getServer().getWorldPath(LevelResource.ROOT);
+            MobHandler.loadMobsSpawnTime(serverLevel, worldPath);
+            GravesLogger.setWorldDirectory(worldPath);
+        }
+    }
+
+    @SubscribeEvent
+    public static void playerLoggedInEvent(PlayerEvent.PlayerLoggedInEvent event) {
+        if (!event.getEntity().level.isClientSide()) {
+            if (GSConfigs.DEBUG_MODE.get()) {
+                LOGGER.info("PlayerEvent.PlayerLoggedInEvent triggered for {}", event.getEntity().getScoreboardName());
+            }
+            MobHandler.setMobSpawnTime(event.getEntity());
+        }
+    }
+
+    @SubscribeEvent
+    public static void playerRespawnInEvent(PlayerEvent.PlayerRespawnEvent event) {
+        if (!event.getEntity().level.isClientSide()) {
+            if (GSConfigs.DEBUG_MODE.get()) {
+                LOGGER.info("PlayerEvent.PlayerRespawnEvent triggered for {}", event.getEntity().getScoreboardName());
+            }
+            MobHandler.setMobSpawnTime(event.getEntity());
+        }
+    }
+
+    @SubscribeEvent
+    public static void onRegisterCommands(RegisterCommandsEvent event) {
+        if (GSConfigs.DEBUG_MODE.get()) {
+            LOGGER.info("RegisterCommandsEvent triggered");
+        }
+
+        var dispatcher = event.getDispatcher();
+        var node = dispatcher.register(GSCommands.root());
+        dispatcher.register(GSCommands.getAlias().redirect(node));
+    }
+
 }
