@@ -1,26 +1,23 @@
 package nightkosh.gravestone.helper;
 
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockFlower;
-import net.minecraft.block.state.IBlockState;
-import net.minecraft.init.Blocks;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
-import net.minecraft.nbt.NBTTagCompound;
-import net.minecraft.util.EnumFacing;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.IBlockAccess;
-import net.minecraft.world.World;
-import net.minecraftforge.common.IPlantable;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.tags.BlockTags;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.FlowerBlock;
+import net.minecraft.world.level.block.state.BlockState;
 import nightkosh.gravestone.api.grave.EnumGraveType;
-import nightkosh.gravestone.block.enums.EnumGraves;
-import nightkosh.gravestone.config.Config;
-import nightkosh.gravestone.core.GSBlock;
-import nightkosh.gravestone.inventory.GraveInventory;
-import nightkosh.gravestone.tileentity.TileEntityGraveStone;
-
-import java.util.Arrays;
-import java.util.List;
+import nightkosh.gravestone.block.BlockGraveStone;
+import nightkosh.gravestone.block_entity.GraveStoneBlockEntity;
+import nightkosh.gravestone.config.GSConfigs;
+import nightkosh.gravestone.core.GSBlocks;
+import nightkosh.gravestone.gui.container.GraveInventory;
+import org.apache.commons.lang3.StringUtils;
 
 /**
  * GraveStone mod
@@ -30,32 +27,31 @@ import java.util.List;
  */
 public class GraveStoneHelper {
 
-
-    public static final List<BlockFlower> FLOWERS = Arrays.asList(Blocks.YELLOW_FLOWER, Blocks.RED_FLOWER);
-
-
     /**
      * Check ground type and replace it on dirt if it grass or mycelium
      */
-    public static void replaceGround(World world, BlockPos pos) {
-        Block botBlock = world.getBlockState(pos).getBlock();
+    public static void replaceGround(Level level, BlockPos pos) {
+        var block = level.getBlockState(pos).getBlock();
 
-        if (botBlock.equals(Blocks.GRASS) || botBlock.equals(Blocks.MYCELIUM)) {
-            world.setBlockState(pos, Blocks.DIRT.getBlockState().getBaseState());
+        if (block.equals(Blocks.GRASS_BLOCK) ||
+                block.equals(Blocks.MYCELIUM) ||
+                block.equals(Blocks.DIRT_PATH) ||
+                block.equals(Blocks.PODZOL)) {
+            level.setBlock(pos, Blocks.DIRT.defaultBlockState(), 3);
+        } else if (block.equals(Blocks.STONE)) {
+            level.setBlock(pos, Blocks.COBBLESTONE.defaultBlockState(), 3);
         }
     }
 
-    public static void spawnMob(World world, BlockPos pos) {
+    public static void spawnMob(Level level, BlockPos pos) {
         //TODO !!!
-//        if (Config.spawnMobAtGraveDestruction && world.rand.nextInt(10) == 0) {
-//            TileEntityGraveStone tileEntity = (TileEntityGraveStone) world.getTileEntity(pos);
-//
-//            if (tileEntity != null) {
-        //TODO
-//                Entity mob = GSMobSpawn.getMobEntity(world, tileEntity.getGraveType(), pos.getX(), pos.getY(), pos.getZ());
+//        if (GSConfigs.spawnMobAtGraveDestruction && level.random.nextInt(10) == 0) {
+//            if (level.getBlockEntity(pos) instanceof GraveStoneBlockEntity grave) {
+////        TODO
+//                var mob = GSMobSpawn.getMobEntity(level, grave.getGraveType(), pos.getX(), pos.getY(), pos.getZ());
 //
 //                if (mob != null) {
-//                    GSMobSpawn.spawnMob(world, mob, pos.getX(), pos.getY(), pos.getZ(), false);
+//                    GSMobSpawn.spawnMob(level, mob, pos.getX(), pos.getY(), pos.getZ(), false);
 //                }
 //            }
 //        }
@@ -64,51 +60,33 @@ public class GraveStoneHelper {
     /**
      * Check can be grave placed on this type of surface
      */
-    public static boolean canPlaceBlockAt(World world, BlockPos pos) {
-        return canPlaceBlockAt(world, world.getBlockState(pos).getBlock(), pos);
+    public static boolean canPlaceBlockAt(BlockState state) {
+        return GSConfigs.CAN_PLACE_GRAVES_EVERY_WHERE.get() ||
+                state.is(BlockTags.MINEABLE_WITH_SHOVEL);
     }
 
-    public static boolean canPlaceBlockAt(World world, Block block, BlockPos pos) {
-        if (Config.canPlaceGravesEveryWhere) {
-            return true;
-        } else {
-            String tool = block.getHarvestTool(world.getBlockState(pos));
-            return tool != null && tool.equals("shovel");
-        }
-    }
-
-    public static void addSwordInfo(NBTTagCompound nbt, ItemStack sword) {
-        NBTTagCompound swordNBT = new NBTTagCompound();
-        sword.writeToNBT(swordNBT);
-        nbt.setTag("Sword", swordNBT);
+    public static void addSwordInfo(CompoundTag tag, ItemStack sword) {
+        var swordTag = new CompoundTag();
+        sword.save(swordTag);
+        tag.put("Sword", swordTag);
     }
 
     public static ItemStack getSwordAsGrave(Item grave, ItemStack sword) {
-        ItemStack graveStoneStack = new ItemStack(grave, 1, EnumGraves.SWORD.ordinal());
-        NBTTagCompound nbt = new NBTTagCompound();
-        nbt.setBoolean("Purified", false);
-        GraveStoneHelper.addSwordInfo(nbt, sword);
-
-        graveStoneStack.setTagCompound(nbt);
+        var graveStoneStack = new ItemStack(grave);
+        var tag = new CompoundTag();
+        tag.putBoolean("Purified", false);
+        GraveStoneHelper.addSwordInfo(tag, sword);
+        graveStoneStack.setTag(tag);
         return graveStoneStack;
     }
 
-    public static boolean canFlowerBePlaced(World world, BlockPos pos, ItemStack itemStack, TileEntityGraveStone te) {
-        if (canFlowerBePlacedOnGrave(te)) {
-            Item item = itemStack.getItem();
-            if (Block.getBlockFromItem(item) instanceof BlockFlower) {
-                return true;
-            } else if (item instanceof IPlantable) {
-                IBlockState soil = world.getBlockState(pos.down());
-                return Block.getBlockFromItem(item).canSustainPlant(soil, world, pos.down(), EnumFacing.UP, (IPlantable) item);
-            }
+    public static boolean canFlowerBePlaced(Level level, BlockPos pos, ItemStack itemStack, GraveStoneBlockEntity entity) {
+        if (Block.byItem(itemStack.getItem()) instanceof FlowerBlock flower) {
+            var posBelow = pos.below();
+            var soil = level.getBlockState(posBelow);
+            return soil.canSustainPlant(level, posBelow, Direction.UP, flower);
         }
         return false;
-    }
-
-    public static boolean canFlowerBePlacedOnGrave(TileEntityGraveStone te) {
-        return !te.isSwordGrave() && (te.getGraveType().getGraveType() == EnumGraveType.VERTICAL_PLATE ||
-                te.getGraveType().getGraveType() == EnumGraveType.CROSS);
     }
 
     public static class RestrictedArea {
@@ -126,11 +104,11 @@ public class GraveStoneHelper {
             this.lastPoint = new BlockPos(endX, endY, endZ);
         }
 
-        public boolean isInArea(World world, BlockPos pos) {
-            return world.provider.getDimension() == dimensionId &&
+        public boolean isInArea(Level level, BlockPos pos) {
+            return false &&//TODO level.provider.getDimension() == dimensionId &&
                     pos.getX() >= firstPoint.getX() && pos.getX() <= lastPoint.getX() &&
-                    pos.getY() >= firstPoint.getY() && pos.getY() <= lastPoint.getY() &&
-                    pos.getZ() >= firstPoint.getZ() && pos.getZ() <= lastPoint.getZ();
+                            pos.getY() >= firstPoint.getY() && pos.getY() <= lastPoint.getY() &&
+                            pos.getZ() >= firstPoint.getZ() && pos.getZ() <= lastPoint.getZ();
         }
 
         public static RestrictedArea getFromString(String area) {
@@ -157,33 +135,20 @@ public class GraveStoneHelper {
         }
     }
 
-    /**
-     * Drop grave as item block
-     */
-    public static void dropBlock(World world, BlockPos pos, IBlockState state) {
-        ItemStack itemStack = getBlockItemStack(world, pos, state);
+    public static void dropBlockWithoutInfo(Level level, GraveStoneBlockEntity graveEntity) {
+        if (GSConfigs.DROP_GRAVE_BLOCK_AT_DESTRUCTION.get()) {
+            if (graveEntity != null) {
+                var grave = ((BlockGraveStone) graveEntity.getBlockState().getBlock());
+                var itemStack = new ItemStack(GSBlocks.getGraveBlock(grave.graveType, grave.material).asItem());
 
-        if (itemStack != null) {
-            GraveInventory.dropItem(itemStack, world, pos);
-        }
-    }
+                if (graveEntity.isSwordGrave()) {
+                    graveEntity.dropSword();
+                } else {
+                    var tag = new CompoundTag();
+                    tag.putBoolean("Purified", true);
 
-    public static void dropBlockWithoutInfo(World world, BlockPos pos, IBlockState state) {
-        if (Config.dropGraveBlockAtDestruction) {
-            ItemStack itemStack = new ItemStack(Item.getItemFromBlock(GSBlock.GRAVE_STONE), 1);
-            TileEntityGraveStone tileEntity = (TileEntityGraveStone) world.getTileEntity(pos);
-
-            if (tileEntity != null) {
-                if (tileEntity.isSwordGrave()) {
-                    tileEntity.dropSword();
-                } else if (itemStack != null) {
-                    NBTTagCompound nbt = new NBTTagCompound();
-                    itemStack.setItemDamage(tileEntity.getGraveTypeNum());
-                    nbt.setBoolean("Mossy", tileEntity.isMossy());
-                    nbt.setBoolean("Purified", true);
-
-                    itemStack.setTagCompound(nbt);
-                    GraveInventory.dropItem(itemStack, world, pos);
+                    itemStack.setTag(tag);
+                    GraveInventory.dropItem(itemStack, level, graveEntity.getBlockPos());
                 }
             }
         }
@@ -192,68 +157,29 @@ public class GraveStoneHelper {
     /**
      * Get grave block as item block
      */
-    public static ItemStack getBlockItemStack(IBlockAccess access, BlockPos pos, IBlockState state) {
-        if (Config.dropGraveBlockAtDestruction) {
-            ItemStack itemStack = new ItemStack(Item.getItemFromBlock(GSBlock.GRAVE_STONE), 1);
-            TileEntityGraveStone tileEntity = (TileEntityGraveStone) access.getTileEntity(pos);
+    public static ItemStack getBlockItemStack(Level level, BlockPos pos, BlockState state) {
+        if (GSConfigs.DROP_GRAVE_BLOCK_AT_DESTRUCTION.get() &&
+                level.getBlockEntity(pos) instanceof GraveStoneBlockEntity grave) {
+            var graveBlock = ((BlockGraveStone) grave.getBlockState().getBlock());
+            var itemStack = new ItemStack(GSBlocks.getGraveBlock(graveBlock.graveType, graveBlock.material).asItem());
 
-            if (tileEntity != null) {
-                NBTTagCompound nbt = new NBTTagCompound();
-                itemStack.setItemDamage(tileEntity.getGraveTypeNum());
-
-                if (tileEntity.getDeathTextComponent().isLocalized()) {
-                    nbt.setBoolean("isLocalized", true);
-                    nbt.setString("name", tileEntity.getDeathTextComponent().getName());
-                    nbt.setString("KillerName", tileEntity.getDeathTextComponent().getKillerName());
-                }
-
-                nbt.setString("DeathText", tileEntity.getDeathTextComponent().getDeathText());
-                nbt.setInteger("Age", tileEntity.getAge());
-
-                if (tileEntity.isSwordGrave()) {
-                    GraveStoneHelper.addSwordInfo(nbt, tileEntity.getSword());
-                }
-
-                nbt.setBoolean("Enchanted", tileEntity.isEnchanted());
-                nbt.setBoolean("Mossy", tileEntity.isMossy());
-
-                nbt.setBoolean("Purified", true);
-
-                itemStack.setTagCompound(nbt);
+            var tag = new CompoundTag();
+            tag.putBoolean("Purified", true);
+            if (grave.getAge() > 0) {
+                tag.putInt("Age", grave.getAge());
             }
 
+            if (StringUtils.isNoneBlank(grave.getDeathMessageJson())) {
+                tag.putString("deathMessageJson", grave.getDeathMessageJson());
+            }
+            if (grave.isSwordGrave()) {
+                GraveStoneHelper.addSwordInfo(tag, grave.getSword());
+            }
+
+            itemStack.setTag(tag);
             return itemStack;
-        } else {
-            return ItemStack.EMPTY;
         }
+        return ItemStack.EMPTY;
     }
 
-//TODO #245
-//    public static void replaceOldGraveByNew(World world, BlockPos pos) {
-//        IBlockState oldState = world.getBlockState(pos);
-//        TileEntity oldTe = world.getTileEntity(pos);
-//        //TODO !!!!!!!!!!!!!!!!!
-//
-//        IBlockState newState = GSBlock.GRAVE_STONE.getStateFromMeta(GSBlock.GRAVE_STONE.getMetaFromState(oldState));//= GSBlock.GRAVE_STONE.getDefaultState();
-//
-//        world.setBlockState(pos, newState);
-//        TileEntity newTe = world.getTileEntity(pos);
-//        if (oldTe  != null && newTe != null && oldTe instanceof TileEntityGraveStone && newTe instanceof TileEntityGraveStone) {
-//            TileEntityGraveStone oldTeGS = (TileEntityGraveStone) oldTe;
-//            TileEntityGraveStone newTeGS = (TileEntityGraveStone) newTe;
-//
-//            newTeGS.setGraveType(oldTeGS.getGraveTypeNum());
-//            newTeGS.setEnchanted(oldTeGS.isEnchanted());
-//            newTeGS.setMossy(oldTeGS.isMossy());
-//            newTeGS.setAge(oldTeGS.getAge());
-//            newTeGS.setDeathTextComponent(oldTeGS.getDeathTextComponent());
-//            newTeGS.setInventory(oldTeGS.getInventory());
-//
-//            newTeGS.setSword(oldTeGS.getSword());
-//            newTeGS.setFlower(oldTeGS.getFlower());
-//            newTeGS.setOwner(oldTeGS.getOwner());
-//            newTeGS.setPurified(oldTeGS.isPurified());
-//            newTeGS.setSpawnerHelper(oldTeGS.getSpawnerHelper());
-//        }
-//    }
 }
