@@ -1,6 +1,12 @@
 package nightkosh.gravestone.capability;
 
-import java.util.*;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.nbt.ListTag;
+import net.minecraft.nbt.Tag;
+import net.minecraftforge.common.util.INBTSerializable;
+
+import java.util.ArrayDeque;
+import java.util.Deque;
 
 /**
  * GraveStone mod
@@ -8,9 +14,11 @@ import java.util.*;
  * @author NightKosh
  * @license Lesser GNU Public License v3 (http://www.gnu.org/licenses/lgpl.html)
  */
-public class Backups implements IBackups {
+public class Backups implements IBackups, INBTSerializable<CompoundTag> {
 
-    private Deque<Backup> backups = new ArrayDeque<>(5);
+    private static final int MAX = 5;
+
+    private Deque<Backup> backups = new ArrayDeque<>(MAX);
 
     @Override
     public Deque<Backup> getBackups() {
@@ -19,22 +27,22 @@ public class Backups implements IBackups {
 
     @Override
     public void setBackups(Deque<Backup> backups) {
-        this.backups = backups;
+        this.backups = backups == null ?
+                new ArrayDeque<>(MAX) :
+                backups;
     }
 
     @Override
     public Backup getBackup(int num) {
         if (num == 0) {
             return backups.getFirst();
-        } else {
-            Iterator<Backup> it = backups.iterator();
-            int i = 0;
-            while (it.hasNext()) {
-                Backup backup = it.next();
+        } else if (num > 0 && num < backups.size()) {
+            var it = backups.iterator();
+            for (int i = 0; it.hasNext(); i++) {
+                var backup = it.next();
                 if (i == num) {
                     return backup;
                 }
-                i++;
             }
         }
         return null;
@@ -42,9 +50,42 @@ public class Backups implements IBackups {
 
     @Override
     public void addBackup(Backup backup) {
-        backups.addFirst(backup);
-        if (backups.size() > 5) {
+        if (backup != null) {
+            backups.addFirst(backup);
+            if (backups.size() > 5) {
+                backups.removeLast();
+            }
+        }
+    }
+
+
+    @Override
+    public CompoundTag serializeNBT() {
+        var tag = new CompoundTag();
+        var list = new ListTag();
+
+        for (var b : backups) {
+            list.add(b.toNBT());
+        }
+
+        tag.put("Backups", list);
+        return tag;
+    }
+
+    @Override
+    public void deserializeNBT(CompoundTag tag) {
+        backups.clear();
+
+        if (!tag.contains("Backups", Tag.TAG_LIST)) return;
+
+        var list = tag.getList("Backups", Tag.TAG_COMPOUND);
+        for (int i = 0; i < list.size(); i++) {
+            backups.addLast(Backup.fromNBT(list.getCompound(i)));
+        }
+
+        while (backups.size() > MAX) {
             backups.removeLast();
         }
     }
+
 }
